requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
cloudinary==1.44.1
supabase==2.3.4
httpx==0.27.0
pydantic==2.11.9
starlette==0.37.2
